Move CreateRegexArray on luet-build
