+++ luet-build/cmd/tree/validate.go
@@ -1,18 +1,7 @@
-// Copyright © 2020 Ettore Di Giacinto <mudler@gentoo.org>
-//                  Daniele Rondina <geaaru@sabayonlinux.org>
-//
-// This program is free software; you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation; either version 2 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License along
-// with this program; if not, see <http://www.gnu.org/licenses/>.
+/*
+Copyright © 2020-2023 Macaroni OS Linux
+See AUTHORS and LICENSE for the license details and contributors.
+*/
 
 package cmd_tree
 
@@ -25,7 +14,7 @@ import (
 	"strconv"
 	"sync"
 
-	helpers "github.com/geaaru/luet/cmd/helpers"
+	helpers "github.com/geaaru/luet/luet-build/cmd/helpers"
 	compiler "github.com/geaaru/luet/pkg/compiler"
 	sd "github.com/geaaru/luet/pkg/compiler/backend"
 	"github.com/geaaru/luet/pkg/compiler/types/options"
